move NameTag to components-temp
